% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{bb_shrink}
\alias{bb_shrink}
\title{Bounding Box Shrink by Factor}
\usage{
bb_shrink(bb, e = 1)
}
\arguments{
\item{bb}{bounding box}

\item{e}{numeric, defaults to 1.
Negative value will grow the bounding box.}
}
\value{
bounding box
}
\description{
Shrinks a bounding box by a factor.
}
\examples{
\dontrun{
# uses example data in the sf package
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
nc_bbox <- sf::st_bbox(nc)
cropped_bbox <- bb_shrink(bb = nc_bbox, e = 0.7)
}
}
