% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kmz.R
\name{collapse_geo}
\alias{collapse_geo}
\title{Collapse Geo}
\usage{
collapse_geo(sfdf, id_var)
}
\arguments{
\item{sfdf}{object of class sf}

\item{id_var}{string naming the column to collapse on}
}
\description{
Many geometries come to us with multiple rows representing
a single unique geometry. This function is designed specifically
with trails in mind, so it collapses to MULTILINESTRING
}
\examples{
\dontrun{
ex_onf_trails <- read_kmz(trailcover_example("onf.kmz"))
ex_onf_simple <- collapse_geo(ex_onf_trails, "Name")
}
}
