% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kmz.R
\name{export_as_kmz}
\alias{export_as_kmz}
\title{Export as KMZ}
\usage{
export_as_kmz(my_sf, fname)
}
\arguments{
\item{my_sf}{object of class sf}

\item{fname}{string file name ending with '.kmz'}
}
\description{
Export as KMZ
}
\examples{
\dontrun{
export_as_kmz(my_tracks, "mytracks.kmz")
}
}
