% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{get_bbox}
\alias{get_bbox}
\title{Get bounding Box}
\usage{
get_bbox(geolist)
}
\arguments{
\item{geolist}{is a single geometry or a list of geometries}
}
\value{
bounding box
}
\description{
Identifies the largest bounding box that covers
all geometries in list
}
\details{
You do not need this unless you want an underlying map.
For example, you are plotting a trail and a track and
want to situate it over roads or other features but still
have the bounding box focused on the trail.
}
