% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_coverage}
\alias{get_coverage}
\title{Get Coverage
creates a sf df of the intersection of
what has been completed and the trail network}
\usage{
get_coverage(big_sf, little_sf)
}
\arguments{
\item{big_sf}{sf trail network map}

\item{little_sf}{sf trail gpx you completed}
}
\description{
Get Coverage
creates a sf df of the intersection of
what has been completed and the trail network
}
