% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_covered_portion}
\alias{get_covered_portion}
\title{Get Covered Portion
creates a dataframe of all trails in the trail
network and identifies how much is completed}
\usage{
get_covered_portion(big_sf, trail_coverage, var = "TRAIL_NAME", tolerance = 1)
}
\arguments{
\item{big_sf}{sf trail network map}

\item{trail_coverage}{sf output of get_coverage()}

\item{var}{string variable indicating trail name}

\item{tolerance}{numeric value you believe indicates
trail completion. defaults to 1, meaning 100\% of official
trail length covered. practical completions may be lower.}
}
\value{
df one row for each row of big_sf with the
portion covered in trail_coverage
}
\description{
Get Covered Portion
creates a dataframe of all trails in the trail
network and identifies how much is completed
}
