% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{get_cropped_to_bb}
\alias{get_cropped_to_bb}
\title{Crop to Bounding Box}
\usage{
get_cropped_to_bb(sfdf, bb)
}
\arguments{
\item{sfdf}{object of class sf}

\item{bb}{bounding box}
}
\value{
cropped sfdf
}
\description{
Crops a simple feature into a bounding box.
}
\details{
If the bb bounding box is larger than sfdf
sfdf is returned unchanged
}
\examples{
\dontrun{
# uses example data in the sf package
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
nc_bbox <- sf::st_bbox(nc)
cropped_bbox <- bb_shrink(bb = nc_bbox, e = 0.7)
cnc <- get_cropped_to_bb(sfdf = nc, bb = cropped_bbox)
ggplot() + geom_sf(data = nc, color = "blue") + geom_sf(data = cnc, fill = "pink")
}
}
