% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trails.R
\name{get_trails}
\alias{get_trails}
\title{Get Trails for a Track}
\usage{
get_trails(track_sf)
}
\arguments{
\item{track_sf}{}
}
\value{
sf
}
\description{
Pulls trails from Open Street Map for
the bounding box of the passed tracks.
}
\details{
subjet to all memory limits of Overpass
queries through osmdata
}
\examples{
\dontrun{
this_track <- read_geo(trailcover_example("Lake_Angeles.gpx"))
this_trails <- get_trails(this_track)
}
}
