% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coverage.R
\name{map_coverage}
\alias{map_coverage}
\title{Map Coverage}
\usage{
map_coverage(
  trail_network,
  covered_sf,
  underlying_map = NA,
  title = "Your Completed Trail Network",
  caption = "created with R/trailcover"
)
}
\arguments{
\item{trail_network}{sf the trail network data
same data as used as big_sf in the other functions.}

\item{covered_sf}{sf output of get_coverage()}

\item{underlying_map}{sf that might represent the
area around your trail network.}

\item{title}{string title you want for your map.
defaults to "Your Completed Trail Network"}

\item{caption}{string caption you want for your map.
defaults to "created with R/trailcover"}
}
\description{
Creates a visual representation of the portion
of trails covered overlaying the trail network
}
\examples{
ex_onp_trails <- read_geo(trailcover_example("onp.geojson"))
this_track <- read_geo(trailcover_example("Lake_Angeles.gpx"))
onp_tracked <- get_coverage(big_sf = ex_onp_trails,
                            little_sf = this_track)
map_coverage(trail_network = ex_onp_trails,
             covered_sf = onp_tracked)
}
