% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_track_v_trail.R
\name{map_track_v_trail}
\alias{map_track_v_trail}
\title{Map Track versus Trails}
\usage{
map_track_v_trail(
  trail_sf,
  track_sf,
  underlying_map = NA,
  trail_sf_name_col = "TRAIL_NAME",
  track_sf_name_col = "name",
  trail_color = "grey70",
  track_color = "deeppink"
)
}
\arguments{
\item{trail_sf}{sf object representing the base layer}

\item{track_sf}{sf object representing the top layer}

\item{underlying_map}{sf object that will make the baseplot}

\item{trail_sf_name_col}{string the name of the column identifying the label}

\item{track_sf_name_col}{string the name of the column identifying the label}

\item{trail_color}{color defaults to 'grey70'}

\item{track_color}{color defaults to 'deeppink'}
}
\value{
ggplot object
}
\description{
Creates a spatial plot overlying a track and a trail.
Allows for an underlying map that can be used to add context.
}
\examples{
\dontrun{

# you want to pass at a minimum two sfs
ex_onp_trails <- read_geo(trailcover_example("onp.geojson"))
this_trail <- ex_onp_trails[ex_onp_trails$TRAIL_NAME == 'LAKE ANGELES TRAIL',]
this_track <- read_geo(trailcover_example("Lake_Angeles.gpx"))
map_track_v_trail(this_trail, this_track)
}
}
