% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onf_trails}
\alias{onf_trails}
\title{Olympic National Forest Trails}
\format{
An object of class \code{SimpleFeaturesDataFrame}
}
\source{
\url{URL to your source data, if applicable}
}
\usage{
onf_trails
}
\description{
A dataset containing geographic feature lines representing trails
in Olympic National Forest
}
\keyword{datasets}
