% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onp_trails}
\alias{onp_trails}
\title{Olympic National Park Trails}
\format{
An object of class \code{SimpleFeaturesDataFrame}
}
\source{
\url{URL to your source data, if applicable}
}
\usage{
onp_trails
}
\description{
A dataset containing geographic feature lines representing trails
in Olympic National Park
}
\keyword{datasets}
