% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_geo.R
\name{read_geo}
\alias{read_geo}
\title{Read Geo}
\usage{
read_geo(fpath)
}
\arguments{
\item{fpath}{}
}
\description{
Attempts to identify if your file needs special handling
like it is a kmz or a gpx, otherwise attempts
to read directly with sf::st_read()
}
\details{
If you want to read in a whole folder of .gpx tracks,
you should use read_gpx_in_folder() instead
}
\examples{
\dontrun{

ex_onf_trails <- read_geo(trailcover_example("onf.kmz"))
}
}
