% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_geo.R
\name{read_gpx_in_folder}
\alias{read_gpx_in_folder}
\title{Read GPX in Folder}
\usage{
read_gpx_in_folder(folder_name, crs = 0)
}
\arguments{
\item{folder_name}{string path to a folder containing .gpx files}

\item{crs}{one of (i) character: a string accepted by GDAL, (ii) integer, a valid EPSG value (numeric), or (iii) an object of class crs}
}
\value{
object of class sf
}
\description{
Reads all of the GPX in a folder into a simple features dataframe
}
\examples{
\dontrun{
my_tracks <- read_gpx_in_folder("path_to_my_tracks")
}
}
