% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kmz.R
\name{read_kmz}
\alias{read_kmz}
\title{Read a Single KMZ into a Simple Features}
\usage{
read_kmz(fpath)
}
\arguments{
\item{fpath}{string path to gpx file}
}
\description{
Read a Single KMZ into a Simple Features
}
\examples{
\dontrun{
ex_onf_trails <- read_kmz(trailcover_example("onf.kmz"))
my_kmz_trails <- read_single_gpx("path_to_my_file.kmz")
}
}
