% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_geo.R
\name{read_single_gpx}
\alias{read_single_gpx}
\title{Read a Single GPX into a Simple Features
Uses track_points and/or tracks layers
Attempts to capture total elevation gain (meters)}
\usage{
read_single_gpx(fpath)
}
\arguments{
\item{fpath}{string path to gpx file}
}
\value{
object of class sf
}
\description{
Read a Single GPX into a Simple Features
Uses track_points and/or tracks layers
Attempts to capture total elevation gain (meters)
}
\examples{
\dontrun{
ex_track <- read_single_gpx(trailcover_example("Lake_Angeles.gpx"))
my_track <- read_single_gpx("path_to_my_track.gpx")
}
}
