% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{trailcover_example}
\alias{trailcover_example}
\title{Get path to example}
\usage{
trailcover_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
trailcover comes bundled an example gpx in \code{inst/extdata}
directory. This function allows access
}
\examples{
trailcover_example()
trailcover_example("Lake_Angeles.gpx")
}
