% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{get_bbox}
\alias{get_bbox}
\title{Get bounding Box}
\usage{
get_bbox(geolist)
}
\arguments{
\item{geolist}{is a single geometry or a list of geometries}
}
\value{
bounding box
}
\description{
Identifies the largest bounding box that covers
all geometries in list
}
