% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_covered_percent}
\alias{get_covered_percent}
\title{get_covered_percent}
\usage{
get_covered_percent(big_sf, covered_portion, var = "TRAIL_NAME")
}
\arguments{
\item{big_sf}{sf trail network map}

\item{covered_portion}{output of get_covered_portion()}

\item{var}{string variable indicating trail name}
}
\description{
get_covered_percent
}
\examples{
\dontrun{
tc <- get_coverage(trail_network, my_gpx_lines)
cp <- get_covered_portion(trail_network,tc, tolerance = 0.95)
get_covered_percent(trail_network, cp)
}
}
