% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coverage.R
\name{map_coverage}
\alias{map_coverage}
\title{Map Coverage}
\usage{
map_coverage(
  underlying_map,
  trail_network,
  trail_coverage,
  trail_coverage_portion,
  title = "Your Completed Trail Network",
  caption = "created with outside_oly"
)
}
\arguments{
\item{underlying_map}{sf that might represent the
area around your trail network. For the examples,
I am using the Olympic Peninsula of Washington to
underly the Oly National Forest and Park trails}

\item{trail_network}{sf the trail network data
same data as used as big_sf in the other functions.
In the examples, this is sf from ONF.geojson and ONP.geojson}

\item{trail_coverage}{sf output of get_coverage()}

\item{trail_coverage_portion}{df has TRAIL_NAME
and portion covered; output of get_covered_portion()}

\item{title}{string title you want for your map.
defaults to "Your Completed Trail Network"}

\item{caption}{string caption you want for your map.
defaults to "created with outside_oly"}
}
\description{
Creates a visual representation of the portion
of trails covered overlaying the trail network
}
